Fix fp8 moe reverse substring (#46265)
* Revert "[DeepSeek V4] Fix MoE converter substring-matching FP8 scale companions (#45930)"

This reverts commit ecc7e32814750cfaae8de7fd02f462d337ae419c.

* [deepseek_v4] fix FP8 save-side substring match + fp32 gate through softmax

* core_model_loading: WeightConverter.reverse_transform now appends a
  (?=\.|$) token-boundary lookahead to each reversed source so they can't
  substring-match longer sibling tokens. Re-fixes #45794 (the FP8Experts
  scale companion `mlp.experts.gate_up_proj_scale_inv` was being routed
  into the weight-only [Chunk, SplitModulelist] ops on save) without
  needing `$` on every model-specific converter — that was #45930's
  approach and broke load via `process_source_pattern`.
* Add `ConversionOps.supports_round_trip` opt-in flag; the on-the-fly
  `quantization_operation` check in reverse_transform now uses it
  instead of a blanket bail-out. Flagged the ops with audited reverse
  pairs (Chunk/Concatenate, MergeModulelist/SplitModulelist, Transpose,
  Conv3d↔Linear, Ernie pair, _IdentityOp, Fp8Quantize/Fp8Dequantize).
* deepseek_v4 compressors (HCA / Indexer / CSA): drop the trailing
  `.to(chunk_gate.dtype)` so the gate stays fp32 from the
  `+ position_bias` through the softmax — matches the reference
  Compressor.forward which upcasts x to fp32 at entry.
* deepseek_v4 config: only set `attention_factor=1.0` when the compress
  group resolves to `rope_type="yarn"`. Suppresses the noisy
  "Unrecognized keys in rope_parameters for rope_type=default" warning
  on the second post_init pass (where yarn params have already been
  consumed).
* Refresh slow integration test expectations to match the fp32-gate
  output.

Verified on the published DeepSeek-V4-Flash checkpoint:
RUN_SLOW=1 pytest tests/models/deepseek_v4 -k Integration  -> 2 passed.

* Apply suggestions from code review


* scope supports_round_trip to quantization ops only

Drop the flag from the regular shape ops (Chunk, Concatenate, MergeModulelist,
SplitModulelist, Transpose, Conv3d↔Linear, _IdentityOp, the Ernie pair) — those
reverse themselves via their own `reverse_op` property, no flag needed. The
flag only gates `WeightConverter.reverse_transform`'s check on the converter's
`quantization_operation`, so today it's just Fp8Quantize / Fp8Dequantize. Other
quant backends (bnb, torchao, eetq, mxfp4, quanto, ...) stay opt-out until
their round-trip is audited.


* Apply suggestions from code review


* Update src/transformers/core_model_loading.py

* style: ruff format


* up

* the actual fix???

* up

* nit comment for futur

* style: drop unused exception variable in Fp8Dequantize._dequantize_one

* up

* style: apply make fix-repo docstring formatting

---------

+++ src/transformers/models/deepseek_v4/modeling_deepseek_v4.py
@@ -412,9 +412,7 @@ class DeepseekV4HCACompressor(nn.Module):
         if chunk_kv.shape[1] > 0:  # there were at least self.compress_rate tokens
             n_windows = chunk_kv.shape[1] // self.compress_rate
             chunk_kv = chunk_kv.view(batch, n_windows, self.compress_rate, -1)
-            chunk_gate = chunk_gate.view(batch, n_windows, self.compress_rate, -1) + self.position_bias.to(
-                chunk_gate.dtype
-            )
+            chunk_gate = chunk_gate.view(batch, n_windows, self.compress_rate, -1) + self.position_bias
             compressed = self.kv_norm(
                 (chunk_kv * chunk_gate.softmax(dim=2, dtype=torch.float32).to(chunk_kv.dtype)).sum(dim=2)
             )
@@ -515,7 +513,7 @@ class DeepseekV4Indexer(nn.Module):
             n_windows = chunk_kv.shape[1] // self.compress_rate
             ratio = self.compress_rate
             chunk_kv = chunk_kv.view(batch, n_windows, ratio, -1)
-            chunk_gate = chunk_gate.view(batch, n_windows, ratio, -1) + self.position_bias.to(chunk_gate.dtype)
+            chunk_gate = chunk_gate.view(batch, n_windows, ratio, -1) + self.position_bias
 
             # Same Ca / Cb overlap layout as the outer CSA compressor, at index_head_dim.
             new_kv = chunk_kv.new_zeros((batch, n_windows, 2 * ratio, self.head_dim))
@@ -633,7 +631,7 @@ class DeepseekV4CSACompressor(nn.Module):
             n_windows = chunk_kv.shape[1] // self.compress_rate
             ratio = self.compress_rate
             chunk_kv = chunk_kv.view(batch, n_windows, ratio, -1)
-            chunk_gate = chunk_gate.view(batch, n_windows, ratio, -1) + self.position_bias.to(chunk_gate.dtype)
+            chunk_gate = chunk_gate.view(batch, n_windows, ratio, -1) + self.position_bias
 
             # Lay out the two series in [B, n_win, 2*ratio, head_dim]: Cb
             # (`[..., head_dim:]`) goes in the second half (current window),
